solucion a la confusion con el archivo app. Migré todo al archivo adminJuegos
