+++ index.html
@@ -352,7 +352,7 @@
     ></script>
     <script src="https://cdn.jsdelivr.net/npm/uuid@latest/dist/umd/uuidv4.min.js"></script>
 
-    <script src="./js/app.js" type="module"></script>
+    <script src="./js/adminJuegos.js" type="module"></script>
     <script src="./js/masInformacion.js" type="module"></script>
   </body>
 </html>
